+++ packages/components/src/components/carousel/carousel.tsx
@@ -32,6 +32,7 @@ import {
   CAROUSEL_ALIGN_HEADERS,
   CAROUSEL_ARIA_ATTRIBUTES,
   CAROUSEL_GRADIENT_COLORS,
+  CAROUSEL_SLIDES_PER_PAGE,
   CAROUSEL_WIDTHS,
   type CarouselAlignControls,
   type CarouselAlignHeader,
@@ -40,6 +41,7 @@ import {
   type CarouselGradientColor,
   type CarouselHeadingSize,
   type CarouselInternationalization,
+  type CarouselSlidesPerPage,
   type CarouselUpdateEventDetail,
   type CarouselWidth,
   getAmountOfPages,
@@ -68,8 +70,7 @@ const propTypes: PropTypes<typeof Carousel> = {
   wrapContent: AllowedTypes.boolean,
   width: AllowedTypes.oneOf<CarouselWidth>(CAROUSEL_WIDTHS),
   slidesPerPage: AllowedTypes.oneOf<ValidatorFunction>([
-    AllowedTypes.breakpoint('number'),
-    AllowedTypes.oneOf(['auto']),
+    AllowedTypes.breakpoint<CarouselSlidesPerPage>(CAROUSEL_SLIDES_PER_PAGE),
   ]),
   gradientColor: AllowedTypes.oneOf<CarouselGradientColor>(CAROUSEL_GRADIENT_COLORS),
   focusOnCenterSlide: AllowedTypes.boolean,
@@ -134,7 +135,7 @@ export class Carousel {
   @Prop() public width?: CarouselWidth = 'basic';
 
   /** Sets the amount of slides visible at the same time. Can be set to `auto` if you want to define different widths per slide via CSS. */
-  @Prop() public slidesPerPage?: BreakpointCustomizable<number> | 'auto' = 1; // eslint-disable-line @typescript-eslint/no-redundant-type-constituents
+  @Prop() public slidesPerPage?: BreakpointCustomizable<CarouselSlidesPerPage> = 1;
 
   /**
    * @deprecated since v3.0.0, will be removed with next major release, use `pagination` instead.
@@ -188,8 +189,8 @@ export class Carousel {
   private paginationEl: HTMLElement;
   private slides: HTMLElement[] = [];
 
-  private get parsedSlidesPerPage(): BreakpointValues<number> | number | 'auto' {
-    return parseJSON(this.slidesPerPage) as BreakpointValues<number> | number | 'auto';
+  private get parsedSlidesPerPage(): BreakpointValues<CarouselSlidesPerPage> | number | 'auto' {
+    return parseJSON(this.slidesPerPage) as BreakpointValues<CarouselSlidesPerPage> | number | 'auto';
   }
 
   private get parsedDisablePagination(): BreakpointValues<boolean> | boolean {
@@ -243,7 +244,6 @@ export class Carousel {
     this.observeSlides(); // initial, adjust aria attributes on slides
     this.splide = new Splide(this.container, {
       start: this.activeSlideIndex,
-      autoWidth: this.parsedSlidesPerPage === 'auto', // https://splidejs.com/guides/auto-width/#auto-width
       arrows: false,
       easing: motionEasingBase,
       focus: this.focusOnCenterSlide ? 'center' : undefined,
@@ -258,7 +258,7 @@ export class Carousel {
       gap: gridGap,
       // TODO: this uses matchMedia internally, since we also use it, there is some redundancy
       breakpoints: getSplideBreakpoints(
-        this.parsedSlidesPerPage as Exclude<BreakpointCustomizable<number> | 'auto', string>
+        this.parsedSlidesPerPage as Exclude<BreakpointCustomizable<CarouselSlidesPerPage> | 'auto', string>
       ), // eslint-disable-line @typescript-eslint/no-redundant-type-constituents
       // https://splidejs.com/guides/i18n/#default-texts
       i18n: parseJSONAttribute(this.intl || {}), // can only be applied initially atm
@@ -451,9 +451,19 @@ export class Carousel {
     this.amountOfPages = getAmountOfPages(
       this.slides.length,
       // round to sanitize floating numbers
-      this.parsedSlidesPerPage === 'auto' ? 1 : Math.round(getCurrentMatchingBreakpointValue(this.parsedSlidesPerPage))
+      getCurrentMatchingBreakpointValue(this.parsedSlidesPerPage) === 'auto'
+        ? 1
+        : Math.round(getCurrentMatchingBreakpointValue(this.parsedSlidesPerPage as number))
     );
     renderPagination(this.paginationEl, this.getPageCount(), this.splide?.index || 0, this.splide);
+
+    // splideJS needs to be refreshed to apply new 'autoWidth' option which is not supported by splideJS breakpoint feature
+    if (this.splide) {
+      getCurrentMatchingBreakpointValue(this.parsedSlidesPerPage) === 'auto'
+        ? (this.splide.options = { autoWidth: true })
+        : false;
+      this.splide.refresh();
+    }
   };
 
   private onNextKeyDown = (e: KeyboardEvent): void => {