+++ packages/components/src/components/select-wrapper/select-wrapper-dropdown/select-wrapper-dropdown-styles.spec.ts
@@ -3,28 +3,22 @@ import { validateCssAndMatchSnapshot } from '../../../../tests/unit/helpers';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
-    ['down', true, 'none', false, false, false, 'light'],
-    ['down', true, 'none', false, false, true, 'light'],
-    ['down', false, 'none', false, false, false, 'light'],
-    ['down', true, 'none', false, true, false, 'light'],
-    ['down', false, 'none', false, true, false, 'light'],
-    ['up', true, 'none', false, false, false, 'light'],
-    ['up', true, 'none', false, false, true, 'light'],
-    ['up', false, 'none', false, false, false, 'light'],
-    ['up', true, 'none', false, true, false, 'light'],
-    ['up', false, 'none', false, true, false, 'light'],
-    ['down', true, 'none', true, false, false, 'light'],
-    ['down', false, 'none', true, false, false, 'light'],
-    ['down', true, 'none', true, true, false, 'light'],
-    ['down', false, 'none', true, true, false, 'light'],
-    ['up', true, 'none', true, false, false, 'light'],
-    ['up', false, 'none', true, false, false, 'light'],
-    ['up', true, 'none', true, true, false, 'light'],
-    ['up', false, 'none', true, true, false, 'light'],
-  ])(
-    'should return correct css for direction: %s, isOpen: %s, state: %s, disabled: %s, filter: %s, isNativePopoverCase: %s and theme: %s',
-    (...args) => {
-      validateCssAndMatchSnapshot(getComponentCss(...args));
-    }
-  );
+    [true, 'none', false, false, 'light'],
+    [true, 'none', false, false, 'dark'],
+    [false, 'none', false, false, 'light'],
+    [true, 'none', false, true, 'light'],
+    [false, 'none', false, true, 'light'],
+    [false, 'none', false, false, 'dark'],
+    [false, 'none', false, true, 'dark'],
+    [true, 'none', true, false, 'light'],
+    [false, 'none', true, false, 'light'],
+    [true, 'none', true, false, 'dark'],
+    [false, 'none', true, true, 'light'],
+    [true, 'none', true, false, 'light'],
+    [false, 'none', true, false, 'dark'],
+    [true, 'none', true, true, 'dark'],
+    [false, 'none', true, true, 'dark'],
+  ])('should return correct css for isOpen: %s, state: %s, disabled: %s, filter: %s, and theme: %s', (...args) => {
+    validateCssAndMatchSnapshot(getComponentCss(...args));
+  });
 });