+++ packages/components/CHANGELOG.md
@@ -19,10 +19,36 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `aria` prop now supports a colon inside of a value, e.g. `aria="{ 'aria-label': 'Always remember: yes you can!' }"`
   ([#3680](https://github.com/porsche-design-system/porsche-design-system/pull/3680))
 
+### [3.27.3] - 2025-03-21
+
+### [3.27.2] - 2025-03-20
+
+### [3.27.2-rc.0] - 2025-03-18
+
 #### Changed
 
-- `Carousel`: Center layout of Carousel on mobile view
-  ([#3765](https://github.com/porsche-design-system/porsche-design-system/pull/3765))
+- `Table`: add border-bottom styles to `Table Head` to support both, empty and missing `Table Body`
+  ([#3788](https://github.com/porsche-design-system/porsche-design-system/pull/3788))
+
+#### Fixed
+
+- `Styles`: change `Gradient` color values from `rgba` to `hsla` to fix Chrome rendering bug
+  ([#3793](https://github.com/porsche-design-system/porsche-design-system/pull/3793))
+
+### [3.27.1] - 2025-03-05
+
+### [3.27.1-rc.0] - 2025-03-05
+
+#### Fixed
+
+- `Carousel`: `auto` value of `slides-per-page` prop is breakpoint customizable
+  ([#3783](https://github.com/porsche-design-system/porsche-design-system/pull/3783))
+- `Select`, `Multi Select`, `Select Wrapper`: remove positioning with native anchor positioning due to a Chrome bug
+  ([#3780](https://github.com/porsche-design-system/porsche-design-system/pull/3780))
+
+### [3.27.0] - 2025-02-28
+
+### [3.27.0-rc.6] - 2025-02-28
 
 #### Added
 
@@ -46,6 +72,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `Select`, `Select Wrapper`:
   - focus outline becomes default focus style when no filter is used
     ([#3754](https://github.com/porsche-design-system/porsche-design-system/pull/3754))
+- `Carousel`: Center layout of Carousel on mobile view
+  ([#3765](https://github.com/porsche-design-system/porsche-design-system/pull/3765))
+- `Carousel`: `auto` value of `slides-per-page` prop is breakpoint customizable
+  ([#3771](https://github.com/porsche-design-system/porsche-design-system/pull/3771))
 
 ### [3.27.0-rc.5] - 2025-02-20